Separate BlockDataManager for future refactor. (#601)
* Complete sqlite related implementations.

* Fix usage of snapshot db and delta db, since they are mostly used as single reader.
* Change to sqlite 0.25 instead of rusqlite.
* Add missing parts and clean up. (Although not 100% cleaned up)

* Refactor BlockDataManager.

* Merge more.
